Fix bug in drupal/etags-get-function-args
With eldoc enabled, this function could move point unexpectedly to the
end of line when point was on the *definition* of anything listed in the
TAGS file.  (This is due to the `find-tag-noselect`, which can return the
already-open buffer for a given tag's file, if one exists).

Adding a `save-excursion` around the `with-current-buffer` fixes this bug.