Avoid "no tags containing" error in hook skeleton
The hook implementation skeleton funcalls `drupal-get-function-args` to
check whether the new function implementation already exists elsewhere
in the project.  When the value of this variable is
`drupal/etags-get-function-args`, it throws an error for non-existent
function names, making it impossible to insert a new implementation.
Adding `ignore-errors` around the call fixes this.